evas gl common - move program vertext attrib arrays init to shadeer init
this removes glDisableVertexAttribArray() entirely and moves
glEnableVertexAttribArray() and glBindAttribLocation() purely to
initial shader load/compile/init because there is no point redoing
this all the time because they are actually bound to the program. we
still have to determine what they point to with
glVertexAttribPointer() as we go but this does cut it down from the
draw path. this doesnt seem to have a measurable effect on nvidia but
on rpi vc4 this dropped vc4's program/shader overhead thats doing
memcmp()'s from 22% of cpu down to 14% of cpu. so the shader
validation/cahe lookups and so on seem to scale by the more we modify
property-wise on a shader. not pretty. seemingly no other drivers
suffer from this that i have seen, but this is a worth speedup of
putting right in now as it can stand on its own.

@optimize