elput: Add API function to set an evas device
Small API function to allow us to set an Evas_Device on an
Elput_Device. This is an internal API and should not need to be used
by app developers.

This is part of a patch set that will allow us to send devices during
events like mouse_move, mouse_down, etc.

@feature

Signed-off-by: Chris Michael <cp.michael@samsung.com>