elput: Add API funtion to set evas device on a given seat
Small API function to allow us to set an Evas_Device on a given
Elput_Seat. This is an internal API and should not be used by
application developers. This patch also adds an event we can raise
when a seat gets added.

This is part of a patch set that will allow us to send devices during
events like mouse_move, mouse_down, etc.

@feature

Signed-off-by: Chris Michael <cp.michael@samsung.com>