mono: Initial support for event arguments.
For example, the event key,down from efl.Input.Interface:

obj.KEY_DOWN = (object sender, efl.input.KEY_DOWN_Args e) => {
    // e.arg is the actual event argument
    Console.Writeline(e.arg.get_key());
}

Events arguments seems to be working but while testing we found that
there are issues regarding the ownership of Strings as return value
from the C side.

Currently the PInvoke declarations return a plain String which causes
the managed code to "own" the string when marshalling it, regardless
of the original owner intent. The correct seems to be to return
IntPtr's from the PInvoke and use Marshal.PtrToString functions
to create managed strings from the given unmanaged pointer.