Revert "elm_genlist: when appending items to the parent, prepend to the parent"
This reverts commit 43d82e567a2d655a089b6ca3f2d913e6ec52f1dc.

I don't understand this commit. It broke the logical order of
items, as the internal list wouldn't match the order on screen.
Other places in the code didn't seem to make this assumption
that parents come after their children. And for sure my recent
fixes require the parent to come before.

This commit was one of the many reasons why odd/even styles
look often wrong.

See T3086